refactor(map): refactor MapManager to use specialized map generation services
- Split monolithic MapManager into multiple specialized services: HeightBasedMapService, IslandMapService, FairIslandsService, and MapAccessService.
- Maintain 100% API compatibility with original MapManager.
- Move map generation logic for fractal, random, fracture, island, and fair island maps into respective services.
- Implement enhanced fair islands generation with retry and validation logic in FairIslandsService.
- Introduce BaseMapGenerationService as abstract base for shared map generation functionality.
- Simplify MapManager by delegating generation and map data access to services.
- Preserve legacy methods in MapManager with deprecation warnings redirecting to new generateMap method.
- Improve code modularity, maintainability, and testability by separating concerns.

References freeciv map generation logic and maintains behavior consistency.
